extract functions from index.js and add stub specs
