+++ functions/helpers/firebaseHelpers.js
@@ -1,6 +1,7 @@
 const admin = require('firebase-admin');
 const {Client, Status} = require('@googlemaps/google-maps-services-js');
 const gmaps = new Client({});
+module.exports = {getPinForAddress, updateCountForRegion, updateCountForCampaign};
 
 const getPinForAddress = (address) => {
     let markerLoc = {lat: 0, lng: 0};
@@ -18,6 +19,22 @@ const getPinForAddress = (address) => {
     }).catch((e) => {
         console.log(e.response.data.error_message);
     });
+}
+
+const updateCountForCampaign = (country, region) => {
+  let dbRef = admin.firestore().collection('campaigns');
+  return dbRef.where('campaignID', '==', campaignID).get().then((snap)=>{
+    if (snap.empty) {
+      throw new Error('could not find campaign with id: ', campaignID);
+    }
+    const doc = snap.docs[0];
+    let count = doc.data().learnerCount + 1;
+    return doc.ref.update({learnerCount: count}).catch((err)=>{
+      console.error(err);
+    })
+  }).catch((err)=>{
+    console.error(err);
+  });
 }
 
 const updateCountForRegion = (country, region) => {
@@ -92,4 +109,95 @@ const updateCountForRegion = (country, region) => {
         });
 }
 
-module.exports = {getPinForAddress, updateCountForRegion};
+function updatePercentFilled(snap, context) {
+  let data = snap.data();
+  const docRef = admin.firestore().collection('donor_master')
+      .doc(data.sourceDonor)
+      .collection('donations').doc(context.params.donationId);
+  return admin.firestore().collection('campaigns')
+      .where('campaignID', '==', data.campaignID).get().then((snap)=>{
+        if (snap.empty) {
+          throw new Error('missing campaign document for ', data.campaignID);
+        }
+        return snap.docs[0].data().costPerLearner;
+      }).then((costPerLearner)=>{
+        const amount = data.amount;
+        const learnerCount = data.learnerCount;
+        return (learnerCount/Math.round(amount / costPerLearner))*100;
+      }).then((percent)=>{
+        return docRef.set({
+          percentFilled: Math.round(percent),
+        }, {merge: true});
+      }).catch((err)=>{
+        console.error(err);
+      });
+}
+
+function getCostPerLearner(campaignID) {
+  return admin.firestore().collection('campaigns')
+      .where('campaignID', '==', campaignID)
+      .get().then((snap)=>{
+        if (snap.empty) {
+          throw new Error('can\'t find campaign with ID: ', campaignID);
+        }
+        return snap.docs[0].data().costPerLearner;
+      }).catch((err)=>{
+        console.error(err);
+      });
+}
+
+function getDonorID(email) {
+  return admin.auth().getUserByEmail(email)
+      .then((user)=>{
+        return user.uid;
+      }).catch((err)=>{
+        if (err.code === 'auth/user-not-found') {
+          console.log('No Donor found for email: ', email);
+          return '';
+        } else throw new Error(err);
+      });
+}
+
+function getDonation(donorID, donationID) {
+  return admin.firestore().collection('donor_master').doc(donorID)
+      .collection('donations').doc(donationID)
+      .get().then((doc)=>{
+        if (!doc.exists) {
+          throw new Error(
+              donorID,
+              ' is missing Donation Document: ',
+              donationID,
+          );
+        }
+        return {id: doc.id, data: doc.data()};
+      }).catch((err)=>{
+        console.error(err);
+      });
+}
+
+function findObjWithProperty(arr, prop, val) {
+  for (let i=0; i < arr.length; i++) {
+    if (arr[i].hasOwnProperty(prop) && arr[i][prop] === val) {
+      return i;
+    }
+  }
+  return -1;
+}
+function updateMasterLearnerCount(country) {
+  const msgRef = firestore.collection('aggregate_data').doc('data');
+  return msgRef.get().then((doc)=>{
+    let count = doc.data().allLearnersCount + 1;
+    let noCountry = doc.data().allLearnersWithDoNotTrack;
+    if (country === 'no-country') {
+      noCountry++;
+    }
+    return msgRef.update({
+      allLearnersCount: count,
+      allLearnersWithDoNotTrack: noCountry,
+    });
+  }).catch((err)=>{
+    console.error(err);
+  });
+}
+
+
