+++ test/functions/checkForDonationEndDate.spec.js
@@ -1,0 +1,5 @@
+const test = require('firebase-functions-test')();
+const sinon = require('sinon');
+const functions = require('firebase-functions');
+const admin = require('firebase-admin');
+const myFunction = require('../../functions/batchManager');
